refactor(mcp): move character tools to character_tools.py
Move GetCharactersTool, CreateCharacterTool, UpdateCharacterTool from
novel_tools.py to character_tools.py. Clean up unused imports.
