+++ backend/mcp_tools/novel_tools.py
@@ -14,9 +14,7 @@ from typing import Literal
 from novels.models import Novel, NovelCreativeProfile
 from chapters.models import Chapter
 from text.utils import count_words
-from characters.models import Character
-from rag.vector_store import vector_store, VectorStoreError
-from .utils import _invalidate_novel_cache, _invalidate_character_cache, _invalidate_chapter_cache
+from .utils import _invalidate_chapter_cache
 
 
 def _build_creative_profile_summary(
@@ -534,322 +532,6 @@ class UpdateCreativeProfileTool(BaseMCPTool):
         )
 
 
-class GetCharactersArgs(BaseModel):
-    mode: Literal["list", "detail", "network"] = Field(default="list", description="查询模式：list=角色列表概览，detail=单角色详细档案，network=关系网络图")
-    character_id: int | None = Field(default=None, description="角色ID（detail模式必填，network模式可选）")
-    search: str | None = Field(default=None, description="角色名搜索（list模式可选）")
-    include_relations: bool = Field(default=True, description="是否包含人物关系网络（list模式）")
-    include_recent_events: bool = Field(default=True, description="是否包含各角色的最近动态（list模式）")
-    include_memory: bool = Field(default=False, description="是否包含语义检索的相关内容片段（detail模式）")
-    include_inactive: bool = Field(default=False, description="是否包含已失效/休眠的关系（network模式）")
-
-class GetCharactersTool(BaseMCPTool):
-    """获取角色信息（列表/详情/关系网络）"""
-
-    name = "get_characters"
-    description = (
-        "获取当前小说的角色信息，支持三种模式："
-        "\n- list: 角色列表概览（含性格标签、关系概要、最近动态），参数: search, include_relations, include_recent_events"
-        "\n- detail: 单角色详细档案，参数: character_id(必填), include_memory(语义检索)"
-        "\n- network: 关系网络图，参数: character_id(可选,有=单角色,无=全局), include_inactive"
-        "\n适用场景：写作前了解角色阵容、深入了解某个角色、查看人物关系网络。"
-    )
-    category = MCPToolCategory.NOVEL_MANAGEMENT
-    args_schema = GetCharactersArgs
-
-    @staticmethod
-    def _extract_personality_summary(personality: dict[str, Any] | None) -> str:
-        if not personality or not isinstance(personality, dict):
-            return ""
-        parts: list[str] = []
-        for key, value in personality.items():
-            text = str(value).strip()
-            if text and len(text) < 200:
-                parts.append(f"{key}：{text}")
-        return "；".join(parts[:5])
-
-    async def _execute(
-        self,
-        args: GetCharactersArgs,
-        *,
-        db: AsyncSession,
-        user_id: int,
-        novel_id: int,
-        **extra,
-    ) -> MCPToolResult:
-        if args.mode == "detail":
-            return await self._execute_detail(db, novel_id, args.character_id, args.include_memory)
-        elif args.mode == "network":
-            return await self._execute_network(db, novel_id, args.character_id, args.include_inactive)
-        else:
-            return await self._execute_list(db, novel_id, args.search, args.include_relations, args.include_recent_events)
-
-    async def _execute_list(self, db, novel_id, search, include_relations, include_recent_events):
-        query = select(Character).where(Character.novel_id == novel_id)
-        if search:
-            query = query.filter(Character.name.contains(search))
-        result = await db.execute(query)
-        characters = result.scalars().all()
-        char_id_map = {c.id: c for c in characters}
-
-        characters_data = [
-            {
-                "id": c.id,
-                "name": c.name,
-                "personality_summary": self._extract_personality_summary(c.personality),
-                "abilities": c.abilities or [],
-            }
-            for c in characters
-        ]
-
-        relations_data = []
-        if include_relations and characters:
-            try:
-                from characters.models import CharacterRelation
-                rel_result = await db.execute(
-                    select(CharacterRelation).where(
-                        CharacterRelation.novel_id == novel_id,
-                        CharacterRelation.status == "active"
-                    )
-                )
-                all_relations = rel_result.scalars().all()
-                for rel in all_relations:
-                    source = char_id_map.get(rel.source_character_id)
-                    target = char_id_map.get(rel.target_character_id)
-                    if source and target:
-                        relations_data.append({
-                            "source_name": source.name,
-                            "target_name": target.name,
-                            "type": rel.relationship_type,
-                            "intensity": rel.intensity,
-                            "status": rel.status,
-                        })
-            except Exception:
-                pass
-
-        recent_events_summary = ""
-        if include_recent_events and characters:
-            try:
-                from timeline.models import TimelineEntry
-                entry_result = await db.execute(
-                    select(TimelineEntry)
-                    .where(TimelineEntry.novel_id == novel_id)
-                    .order_by(TimelineEntry.updated_at.desc())
-                    .limit(10)
-                )
-                recent_entries = entry_result.scalars().all()
-                if recent_entries:
-                    event_lines = []
-                    for entry in recent_entries:
-                        status_label = {"pending": "待办", "active": "进行中", "completed": "已完成", "resolved": "已回收"}.get(entry.status, entry.status)
-                        line = f"[{entry.category}/{status_label}] {entry.title}"
-                        if entry.description:
-                            line += f" — {entry.description[:50]}"
-                        event_lines.append(line)
-                    recent_events_summary = "\n".join(event_lines)
-                else:
-                    recent_events_summary = "暂无追踪记录。"
-            except Exception:
-                recent_events_summary = ""
-
-        return MCPToolResult(
-            success=True,
-            data={
-                "characters": characters_data,
-                "relations": relations_data,
-                "recent_events_summary": recent_events_summary,
-                "total_characters": len(characters),
-            },
-            metadata={"tool": self.name, "novel_id": novel_id, "mode": "list"}
-        )
-
-    async def _execute_detail(self, db, novel_id, character_id, include_memory):
-        if not character_id:
-            return MCPToolResult(success=False, error="detail 模式需要 character_id")
-
-        result = await db.execute(
-            select(Character)
-            .options(selectinload(Character.novel))
-            .where(Character.id == character_id)
-        )
-        character = result.scalar_one_or_none()
-        if not character:
-            return MCPToolResult(success=False, error=f"角色不存在: {character_id}")
-        if character.novel_id != novel_id:
-            return MCPToolResult(success=False, error=f"角色 {character_id} 不属于当前小说")
-
-        data = {
-            "id": character.id,
-            "novel_id": character.novel_id,
-            "name": character.name,
-            "personality": character.personality,
-            "abilities": character.abilities,
-            "relationships": character.relationships,
-            "created_at": character.created_at.isoformat() if character.created_at else None,
-            "novel": {"id": character.novel.id, "title": character.novel.title} if character.novel else None,
-        }
-
-        if include_memory:
-            try:
-                search_results = await vector_store.search(novel_id=novel_id, query=character.name, top_k=5)
-                data["relevant_content"] = [
-                    {"content": r["content"][:200] + "..." if len(r["content"]) > 200 else r["content"],
-                     "chapter_id": r["metadata"].get("chapter_id")}
-                    for r in search_results
-                ]
-            except VectorStoreError:
-                data["relevant_content"] = []
-
-        return MCPToolResult(
-            success=True,
-            data=data,
-            metadata={"tool": self.name, "novel_id": novel_id, "character_id": character_id, "mode": "detail"}
-        )
-
-    async def _execute_network(self, db, novel_id, character_id, include_inactive):
-        from characters.service import CharacterService
-        service = CharacterService(db, novel_id)
-
-        if character_id:
-            relationships = await service.get_character_relationships(
-                character_id=character_id, include_inactive=include_inactive
-            )
-            return MCPToolResult(
-                success=True,
-                data={"relationships": relationships, "total": len(relationships), "character_id": character_id},
-                metadata={"tool": self.name, "novel_id": novel_id, "mode": "network"}
-            )
-        else:
-            network_data = await service.get_network()
-            return MCPToolResult(
-                success=True,
-                data=network_data,
-                metadata={"tool": self.name, "novel_id": novel_id, "mode": "network"}
-            )
-
-
-class CreateCharacterArgs(BaseModel):
-    name: str = Field(description="角色名称（必填）")
-    personality: dict | None = Field(default=None, description="角色性格/设定字典，建议包含: role(定位), traits(性格), background(背景), motivation(动机), appearance(外貌)")
-    abilities: list[str] | None = Field(default=None, description="角色能力/技能列表")
-
-class CreateCharacterTool(BaseMCPTool):
-    """创建新角色"""
-
-    name = "create_character"
-    description = (
-        "为当前小说创建一个新角色。"
-        "\n适用场景：用户要求添加新角色、AI写作时发现需要新角色、规划角色阵容时。"
-        "\n创建后可通过 get_characters(mode=\"detail\") 查看详情，通过 update_character 修改设定。"
-        "\n注意：name 为必填；personality 建议包含 role(角色定位)、traits(性格特点)、background(背景) 等字段。"
-    )
-    category = MCPToolCategory.NOVEL_MANAGEMENT
-    args_schema = CreateCharacterArgs
-
-    async def _execute(
-        self,
-        args: CreateCharacterArgs,
-        *,
-        db: AsyncSession,
-        user_id: int,
-        novel_id: int,
-        **extra,
-    ) -> MCPToolResult:
-        try:
-            from characters.models import Character
-            character = Character(
-                novel_id=novel_id,
-                name=args.name,
-                personality=args.personality or {},
-                abilities=args.abilities or [],
-            )
-            db.add(character)
-            await db.commit()
-            await db.refresh(character)
-
-            await _invalidate_character_cache(novel_id)
-
-            return MCPToolResult(
-                success=True,
-                data={
-                    "id": character.id,
-                    "name": character.name,
-                    "novel_id": character.novel_id,
-                    "personality": character.personality,
-                    "abilities": character.abilities,
-                },
-                metadata={"tool": self.name, "novel_id": novel_id, "character_id": character.id}
-            )
-        except Exception as e:
-            return MCPToolResult(success=False, error=f"创建角色失败: {str(e)}")
-
-
-class UpdateCharacterArgs(BaseModel):
-    character_id: int = Field(description="角色ID（必填）")
-    name: str | None = Field(default=None, description="新的名称")
-    personality: dict | None = Field(default=None, description="新的性格/设定字典（完全替换旧的）")
-    abilities: list[str] | None = Field(default=None, description="新的能力列表（完全替换旧的）")
-
-class UpdateCharacterTool(BaseMCPTool):
-    """更新角色信息"""
-
-    name = "update_character"
-    description = (
-        "更新已有角色的设定信息。"
-        "\n适用场景：用户要求修改角色设定、AI写作中发现需要调整角色属性时。"
-        "\n只需传入要修改的字段，未传入的字段保持不变。"
-        "\n修改后相关缓存会自动失效，下次查询获取最新数据。"
-    )
-    category = MCPToolCategory.NOVEL_MANAGEMENT
-    args_schema = UpdateCharacterArgs
-
-    async def _execute(
-        self,
-        args: UpdateCharacterArgs,
-        *,
-        db: AsyncSession,
-        user_id: int,
-        novel_id: int,
-        **extra,
-    ) -> MCPToolResult:
-        try:
-            from characters.models import Character
-            result = await db.execute(
-                select(Character).where(Character.id == args.character_id)
-            )
-            character = result.scalar_one_or_none()
-            if not character:
-                return MCPToolResult(success=False, error=f"角色 {args.character_id} 不存在")
-            if character.novel_id != novel_id:
-                return MCPToolResult(success=False, error=f"角色不属于当前小说")
-
-            if args.name is not None:
-                character.name = args.name
-            if args.personality is not None:
-                character.personality = args.personality
-            if args.abilities is not None:
-                character.abilities = args.abilities
-
-            await db.commit()
-            await db.refresh(character)
-
-            await _invalidate_character_cache(novel_id, args.character_id)
-
-            return MCPToolResult(
-                success=True,
-                data={
-                    "id": character.id,
-                    "name": character.name,
-                    "novel_id": character.novel_id,
-                    "personality": character.personality,
-                    "abilities": character.abilities,
-                },
-                metadata={"tool": self.name, "novel_id": novel_id, "character_id": args.character_id}
-            )
-        except Exception as e:
-            return MCPToolResult(success=False, error=f"更新角色失败: {str(e)}")
-
-
 class CreateNewChapterArgs(BaseModel):
     chapter_number: int | None = Field(default=None, description="章节号")
     title: str | None = Field(default=None, description="章节标题（可选）")
@@ -948,7 +630,4 @@ class NovelManagementTools:
         registry.register(GetChapterContentTool())
         registry.register(GetCreativeProfileTool())
         registry.register(UpdateCreativeProfileTool())
-        registry.register(GetCharactersTool())
-        registry.register(CreateCharacterTool())
-        registry.register(UpdateCharacterTool())
         registry.register(CreateNewChapterTool())
